modified timer and rcc files

+++ Inc/rcc/rcc.hpp
@@ -8,7 +8,7 @@ namespace rcc
 
 // temporarily using sysclk variable as 16MHz. Later using function to determine sysclk and bus clocks
 const std::uint32_t SYSCLK = 16000000;
-const std::uint32_t SYSCLK_MHz = 16;
+const std::uint8_t SYSCLK_MHz = 16;
 
 void enableClock_GPIO(GPIO_TypeDef*);
